daily_pc_activity_report/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-pc-activity-report"
version = "1.0.0"
description = "自动生成每日电脑活动简报的工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Daily PC Activity Report"}
]
keywords = ["automation", "report", "file-system", "activity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyYAML>=6.0.1",
]

[project.scripts]
daily-report = "daily_pc_activity_report.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
daily_pc_activity_report = ["py.typed"]
